scripts/requirements.txt
torch-geometric
torchvision==0.12.0